Towards adding unit tests for pdgemm wrapper in COSMA.
