+++ tests/cosma_pdgemm_run.hpp
@@ -56,6 +56,7 @@ bool validate_results(std::vector<double>& v1, std::vector<double>& v2, double e
 // a vector of timings (in milliseconds) of size n_rep
 bool test_pdgemm(int m, int n, int k, // matrix sizes
         int bm, int bn, int bk, // blocks sizes
+        char trans_a, char trans_b, // transpose flags
         int p, int q, // processor grid
         int rank, MPI_Comm comm) {
 
@@ -71,11 +72,6 @@ bool test_pdgemm(int m, int n, int k, // matrix sizes
     // ***********************************
     //   describe the problem parameters
     // ***********************************
-    // transpose flags
-    char trans_a = 'N';
-    char trans_b = 'N';
-    char trans_c = 'N';
-
     double alpha = 1.0;
     double beta = 0.0;
 
@@ -163,67 +159,4 @@ bool test_pdgemm(int m, int n, int k, // matrix sizes
     blacs::Cblacs_gridexit(ctxt);
 
     return validate_results(c_cosma, c_scalapack);
-}
-
-int main(int argc, char **argv) {
-    // **************************************
-    //   setup MPI and command-line parser
-    // **************************************
-    options::initialize(argc, argv);
-
-    MPI_Init(&argc, &argv);
-
-    int rank, P;
-    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-    MPI_Comm_size(MPI_COMM_WORLD, &P);
-
-    // **************************************
-    //   readout the command line arguments
-    // **************************************
-    // matrix dimensions
-    // dim(A) = mxk, dim(B) = kxn, dim(C) = mxn
-    auto m = options::next_int("-m", "--m_dim", "number of rows of A and C.", 1000);
-    auto n = options::next_int("-n", "--n_dim", "number of columns of B and C.", 1000);
-    auto k = options::next_int("-k", "--k_dim", "number of columns of A and rows of B.", 1000);
-
-    // block sizes
-    auto bm = options::next_int("-bm", "--m_block", "block size for the number of rows of A and C.", 128);
-    auto bn = options::next_int("-bn", "--n_block", "block size for the number of columns of B and C.", 128);
-    auto bk = options::next_int("-bk", "--k_block", "block size for the number of columns of A and rows of B.", 128);
-
-    // processor grid decomposition
-    auto p = options::next_int("-p", "--p_row", "number of rows in a processor grid.", 1);
-    auto q = options::next_int("-q", "--q_row", "number of columns in a processor grid.", P);
-
-    if (p * q != P) {
-        std::runtime_error("Number of processors in a grid has to match the number of available ranks.");
-    }
-
-    // **************************************
-    //    output the problem description
-    // **************************************
-    if (rank == 0) {
-        std::cout << "Running PDGEMM on the following problem size:" << std::endl;
-        std::cout << "Matrix sizes: (m, n, k) = (" << m << ", " << n << ", " << k << ")" << std::endl;;
-        std::cout << "Block sizes: (bm, bn, bk) = (" << bm << ", " << bn << ", " << bk << ")" << std::endl;;
-        std::cout << "Processor grid: (prows, pcols) = (" << p << ", " << q << ")" << std::endl;;
-    }
-
-    // *******************************
-    //   multiply and validate
-    // *******************************
-    bool ok = test_pdgemm(m, n, k, bm, bn, bk, p, q, rank, MPI_COMM_WORLD);
-    int result = ok ? 0 : 1;
-    int global_result = 0;
-
-    MPI_Reduce(&result, &global_result, 1, MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
-
-    if (rank == 0) {
-        std::string yes_no = global_result == 0 ? "" : " NOT";
-        std::cout << "Result is" << yes_no << " CORRECT!" << std::endl;
-    }
-
-    MPI_Finalize();
-
-    return 0;
 }
