+++ tests/test_pdgemm.cpp
@@ -1,4 +1,4 @@
-#include <cosma_run.hpp>
+#include <cosma_pdgemm_run.hpp>
 
 #include <gtest/gtest.h>
 #include <gtest_mpi/gtest_mpi.hpp>
