+++ tests/test_strategy.cpp
